feat(changelog): add automation scripts for changelog generation and API documentation
- Implement `generate-changelog.sh` for automated changelog creation from git commits.
- Add `extract-breaking-changes.py` to extract breaking changes from cargo-semver-checks output.
- Create `generate-migration-examples.py` for generating migration examples for API changes.
- Develop `extract-version-annotations.py` to extract "Since" version annotations from Rust doc comments.
- Introduce `generate-compatibility-table.py` to generate API compatibility tables across versions.
- Add `add-phase2-enforcement-batch.py` for batch processing of documentation enforcement in Phase 2 service crates.
- Implement `find-phase2-additional-crates.py` to identify additional high-value crates for Phase 2 completion.
- Create `find-phase2-service-crates.py` to find service crates corresponding to Phase 1 core crates with enforcement.

+++ service/benches/expression_benchmarks.rs
@@ -127,7 +127,7 @@ fn bench_rule_engine(c: &mut Criterion) {
             }),
             else_conditions: None,
             deactivated: Some(false),
-            priority: Some(i as i32),
+            priority: Some(i),
         };
         entity_class.rules.push(rule);
     }
